@@ -9,7 +9,7 @@ git_source(:github) do |repo_name|
   "https://github.com/#{repo_name}.git"
 end
 
-gem 'rails', '~> 7.0.8'
+gem 'rails', '~> 7.1.3'
 gem 'activerecord-session_store'
 gem 'activeresource', '~> 6.0'
 
@@ -24,7 +24,7 @@ gem 'simple_form'
 gem 'nokogiri', '~> 1.16'
 gem 'rdiscount'
 gem 'loofah'
-gem 'rails-html-sanitizer', '~> 1.4.4'
+gem 'rails-html-sanitizer', '~> 1.6'
 gem 'font-awesome-rails'
 gem 'non-stupid-digest-assets'
 gem 'http_accept_language'
@@ -92,7 +92,7 @@ group :test, :development do
   gem 'pry-rails'
   gem 'pry-remote'
   gem 'pry-stack_explorer'
-  gem 'rspec-rails'
+  gem 'rspec-rails', '~> 6.1.2'
   gem 'watchr'
 
   gem 'rubocop', '~> 1.0',    require: false
@@ -104,17 +104,17 @@ group :test, :development do
 end
 
 group :test do
-  gem 'capybara'
+  gem 'capybara', '~> 3.40.0'
   gem 'database_cleaner'
   gem 'factory_bot_rails', '~> 4.8'
   gem 'launchy'
   gem 'rails-controller-testing'
   gem 'rspec_junit_formatter'
-  gem 'selenium-webdriver'
+  gem 'selenium-webdriver', '~> 4.10'
   gem 'shoulda-matchers', require: false
   gem 'simplecov'
   gem 'vcr', '~> 6.0'
-  gem 'webdrivers'
+  gem 'webdrivers', '~> 5.3.0'
   gem 'webmock', '>= 3.5'
 end
 
