@@ -3,11 +3,11 @@
 class MultiYearChartsController < ApplicationController
   layout 'multi_year_charts'
 
-  load_resource only: %i[discard undiscard]
+  load_resource only: %i[show discard undiscard]
 
   before_action :ensure_valid_config
 
-  before_action except: %i[discard undiscard] do
+  before_action except: %i[show discard undiscard] do
     authenticate_user!(show_as: :sign_in)
   end
 
@@ -27,24 +27,35 @@ class MultiYearChartsController < ApplicationController
     end
   end
 
+  def new
+    @multi_year_chart = current_user.multi_year_charts.build(interpolation: false)
+
+    respond_to do |format|
+      format.html { render layout: 'application' }
+    end
+  end
+
   # Part of the My Scenarios view, lists all MYC that are not discarded
   #
   # GET multi_year_charts/list
   def list
-    @multi_year_charts = user_multi_year_charts
+    @multi_year_charts = user_collections
       .kept
       .includes(:user)
-      .order('created_at DESC')
-      .page(params[:page])
-      .per(50)
 
     respond_to do |format|
-      format.html {render :layout => 'application'}
+      format.html { render layout: 'application' }
+    end
+  end
+
+  def show
+    respond_to do |format|
+      format.html { render layout: 'application' }
     end
   end
 
   # Creates a new MultiYearChart record based on the scenario specified in the
-  # params.
+  # params. This is the interpolation route
   #
   # Redirects to the external MYC app when successful.
   #
@@ -65,6 +76,27 @@ class MultiYearChartsController < ApplicationController
       @multi_year_charts = user_multi_year_charts
 
       render :index, status: :unprocessable_entity
+    end
+  end
+
+  def create_collection
+    saved_scenario_ids = create_collection_params.delete(:saved_scenarios)
+
+    collection = current_user.multi_year_charts.build(
+      title: collection_title,
+      interpolation: false
+    )
+
+    saved_scenario_ids.uniq.reject(&:empty?).each do |saved_scenario_id|
+      collection.multi_year_chart_saved_scenarios.build(saved_scenario_id:)
+    end
+
+    if collection.valid?
+      collection.save
+      redirect_to show_multi_year_chart_path(collection)
+    else
+      flash[:error] = t('multi_year_charts.failure')
+      redirect_to list_multi_year_charts_path
     end
   end
 
@@ -137,10 +169,29 @@ class MultiYearChartsController < ApplicationController
       .per(50)
   end
 
+  def user_collections
+    return [] unless current_user
+
+    current_user
+      .multi_year_charts
+      .order(created_at: :desc)
+      .page(params[:page])
+      .per(8)
+  end
+
   def ensure_valid_config
     return if Settings.multi_year_charts_url
 
     redirect_to root_path,
       notice: 'Missing multi_year_charts_url setting in config.yml'
+  end
+
+  def create_collection_params
+    params.require(:multi_year_chart).permit(:title, saved_scenarios: [])
+  end
+
+  def collection_title
+    title = create_collection_params[:title]
+    title.empty? ? t('multi_year_charts.no_title') : title
   end
 end
