@@ -4,7 +4,7 @@
 class SavedScenarioHistoryController < ApplicationController
   before_action :assign_saved_scenario
 
-  before_action only: %i[index update edit] do
+  before_action only: %i[index update] do
     authorize!(:update, @saved_scenario)
   end
 
