@@ -11,18 +11,21 @@ module API
         optional(:title).filled(:string)
         optional(:area_code).filled(:string)
         optional(:end_year).filled(:integer)
-        optional(:scenario_ids).filled(min_size?: 1, max_size?: 10).each(:integer, gt?: 0)
+        optional(:scenario_ids).filled(min_size?: 1, max_size?: 100).each(:integer, gt?: 0)
+        optional(:saved_scenario_ids).filled(min_size?: 1, max_size?: 100).each(:integer, gt?: 0)
       end
     end
 
     def call(transition_path:, params:)
       params = yield validate(params)
       scenario_ids = params.delete(:scenario_ids)
+      saved_scenario_ids = params.delete(:saved_scenario_ids)
 
       transition_path.attributes = params
 
       MultiYearChart.transaction do
         update_scenarios(transition_path, scenario_ids.uniq) if scenario_ids&.any?
+        update_saved_scenarios(transition_path, saved_scenario_ids.uniq) if saved_scenario_ids&.any?
         transition_path.save!
       rescue ActiveRecord::RecordInvalid
         return Failure(transition_path.errors)
@@ -45,6 +48,15 @@ module API
 
       transition_path.scenarios.delete_by(scenario_id: delete_ids)
       new_ids.each { |id| transition_path.scenarios.create!({ scenario_id: id }) }
+    end
+
+    def update_saved_scenarios(transition_path, saved_scenario_ids)
+      existing_ids = transition_path.saved_scenarios.pluck(:saved_scenario_id)
+      new_ids      = saved_scenario_ids - existing_ids
+      delete_ids   = existing_ids - saved_scenario_ids
+
+      transition_path.multi_year_chart_saved_scenarios.delete_by(saved_scenario_id: delete_ids)
+      new_ids.each { |id| transition_path.multi_year_chart_saved_scenarios.create!({ saved_scenario_id: id }) }
     end
   end
 end
