@@ -47,7 +47,7 @@ RSpec.describe API::CreateTransitionPath do
         title: ['is missing'],
         area_code: ['is missing'],
         end_year: ['is missing'],
-        scenario_ids: ['is missing']
+        scenario_ids: ['at least one scenario_id or saved_scenario_id should be present']
       })
     end
   end
@@ -88,9 +88,9 @@ RSpec.describe API::CreateTransitionPath do
     end
   end
 
-  context 'when given more than 10 scenario IDs' do
+  context 'when given more than 100 scenario IDs' do
     let(:params) do
-      super().merge(scenario_ids: (1..11).to_a)
+      super().merge(scenario_ids: (1..101).to_a)
     end
 
     it 'returns a failure' do
@@ -99,7 +99,7 @@ RSpec.describe API::CreateTransitionPath do
 
     it 'returns the errors' do
       expect(result.failure).to eq({
-        scenario_ids: ['size cannot be greater than 10']
+        scenario_ids: ['size cannot be greater than 100']
       })
     end
   end
